Populate InnerException from Python's __cause__ when available
This change only affects some exception handling, but not all of it. Instead, parameterless PythonException constructor is made Obsolete. PythonException.FromPyErr() should be called instead.

FromPyErr will check __cause__, and, if present, parse it into InnerException property.

This change will cause a slowdown, as I found no way to read __cause__ on demand, because InnerException is not virtual.

This addresses #893