fixed deallocation of objects after runtime restart
ObjectOffset.magic method, that was used to determine the offset to GCHandle was comparing object
metatype to Runtime.CLRMetaType, which has changed due to runtime restart.

Overall, ObjectOffset.magic had too much responsibility and relatively unclear behavior.

It was replaced with one of:
- GetDefaultGCHandleOffset, which returns GCHandle offset for types deriving from regular Python types
- ReflectedObjectGCHandle, that returns GCHandle offset for reflected CLR object instances
- InstanceGCHandle, that does the same based on reflected object type rather than instance

The way tp_basicsize and clr_gchandle_offset are calculated is made more robust.
It now requires an explicit base type to be specified, and counts derived type size and GCHandle offset from it.